Switch to Manifold sources as public interface
The library implementation relies heavily on core.async. Therefore it
felt initially like a good idea to expose the upsert/delete streams from
the public API also as core.async channels. But after having read the
rational for the Manifold library
(https://github.com/ztellman/manifold/blob/master/docs/rationale.md) I
got convinced that Manifold is a better abstraction to expose as it
doesn't lock consumers to use core.async.

Furthermore, the plan is to next work on exposing lag and throughput as
periodically querieable values (a stream probably). For this work the
Narrator library looks like a likely choice. This means Manifold will be
a dependency in any case.